v23/factories: fix configuring permission flags
Summary:
D26289 updated Vanadium to 0.1.2 to support go 1.12. However, the
upgrade also introduced a change that separated the permissions flags
and required it to be explicitly configured. This fixes the grail v23
runtime factory so that the permission flags are registered as they
are needed by many grail programs for access control.

Ref T5354
Ref https://github.com/vanadium/core/commit/161f5735b

Test Plan:
- Start the LIMS server locally with both `-v23.permissions.literal` and
  `-v23.permissions.file`.
- `go test -run TestLimsServerLogs$`

Reviewers: ysaito, marius, sbagaria, aeiser

Reviewed By: sbagaria

Subscribers: schandra, mfilippova, alice, cbain

Maniphest Tasks: T5354

Differential Revision: https://phabricator.grailbio.com/D26341

fbshipit-source-id: 4bf35fd