update v.io to v0.1.5 (#1)
Update dependencies here too to cut off the tail of transitive dependencies on older versions.

That is, this package, albeit with one file, is pulling in the older v.io which is in turn pulling in the older dependencies. Cleaning this up should cut off the tail of those older dependencies.